backend/requirements.txt
python-dotenv~=0.17.0
pymongo~=3.11.3
flask-restful
flask~=1.1.2
networkx
advertools~=0.11.1
sqlalchemy
bokeh
keras~=2.4.3
protobuf~=3.8.0
nltk~=3.6.2
numpy~=1.19.0
tweepy~=3.10.0
pandas~=0.25.3
celery~=5.0.5
flask-cors
redis~=3.5.3
